+++ Libplanet/Store/Trie/Nodes/BaseNode.cs
@@ -1,18 +1,10 @@
 #nullable enable
-using Bencodex;
 using Bencodex.Types;
 
 namespace Libplanet.Store.Trie.Nodes
 {
     internal abstract class BaseNode : INode
     {
-        private static Codec _codec;
-
-        static BaseNode()
-        {
-            _codec = new Codec();
-        }
-
         protected BaseNode(INode? value)
         {
             Value = value;
@@ -20,8 +12,6 @@ namespace Libplanet.Store.Trie.Nodes
 
         // It will not support embedded node.
         public INode? Value { get; }
-
-        public byte[] Serialize() => _codec.Encode(ToBencodex());
 
         /// <inheritdoc cref="INode.ToBencodex()"/>
         public abstract IValue ToBencodex();