+++ CHANGES.md
@@ -17,10 +17,31 @@ To be released.  The solution now can be built on Apple Silicon
 
 ### Added APIs
 
+ -  Added `PreEvaluationBlock<T>.DetermineStateRootHash(BlockChain<T>,
+    StateCompleterSet<T>, out IImmutableDictionary<string, IValue>)` overloaded
+    method.  [[#1636]]
+ -  Added `PreEvaluationBlock<T>.DetermineStateRootHash(IAction?, IStateStore,
+    out IImmutableDictionary<string, IValue>)` overloaded method.  [[#1636]]
  -  Parameter `except` for `KBucket.GetRandomPeer()` now defaults to `null`.
     [[#1631]]
+ -  Added `ArrayEqualityComparer<T>` class.  [[#1636]]
 
 ### Behavioral changes
+
+ -  States became to take up much less space than before by reducing unnecessary
+    duplicate data.  Although this guarantees the backward compatibility with
+    the existing state store, in order to take complete effect of this
+    optimization, please replay your existing blockchain from the genesis block
+    with the empty store.  [[#1636]]
+ -  (Libplanet.Analyzers) Rule LAA1002 no more warns about enumerating sorted
+    sets/dictionaries:
+     -  `System.Collections.Generic.SortedDictionary<TKey, TValue>`
+     -  `System.Collections.Generic.SortedSet<T>`
+     -  `System.Collections.Immutable.ImmutableSortedDictionary<TKey, TValue>`
+     -  `System.Collections.Immutable.ImmutableSortedSet<T>`
+     -  `Bencodex.Types.Dictionary` (which became to guarantee enumeration order
+        since [Bencodex 0.4.0-dev.20211116020419+abea0858][Bencodex
+        0.4.0-dev.20211116020419])
 
 ### Bug fixes
 
@@ -29,6 +50,9 @@ To be released.  The solution now can be built on Apple Silicon
 
 ### Dependencies
 
+ -  Upgraded *Bencodex* from 0.4.0-dev.20211123080042+d7f6c810 to
+    [0.4.0-dev.20211205152306+db8d4e1b][Bencodex 0.4.0-dev.20211205152306].
+    [[#1636]]
  -  Upgraded *Planetarium.RocksDbSharp* from 6.2.3 to
     [6.2.4-planetarium][Planetarium.RocksDbSharp 6.2.4-planetarium].
     [[#1635]]
@@ -37,6 +61,9 @@ To be released.  The solution now can be built on Apple Silicon
 
 [#1631]: https://github.com/planetarium/libplanet/pull/1631
 [#1635]: https://github.com/planetarium/libplanet/pull/1635
+[#1636]: https://github.com/planetarium/libplanet/pull/1636
+[Bencodex 0.4.0-dev.20211116020419]: https://www.nuget.org/packages/Bencodex/0.4.0-dev.20211116020419
+[Bencodex 0.4.0-dev.20211205152306]: https://www.nuget.org/packages/Bencodex/0.4.0-dev.20211205152306
 [Planetarium.RocksDbSharp 6.2.4-planetarium]: https://www.nuget.org/packages/Planetarium.RocksDbSharp/6.2.4-planetarium